Split heat pump details into Temps and System screens
- Create heatpump_temps_screen with large fonts (24/32pt) for temperature readings
- Create heatpump_system_screen with sections for compressor, electrical, pressures, EEV, setpoints
- Replace single Details button with three buttons: Controls, Temps, System
- Add color coding for temperature values (warm/cold)
- Add status coloring for compressor/fan (green when active)
- Delete old heatpump_detail_screen (replaced by new split screens)
- Increase button fonts from 16pt to 24pt for better readability
- Fix container height to fill screen properly, buttons at bottom
